top animes showing as expected when enter in search screen

+++ app/src/main/java/com/example/myanimelist/domain/model2/Data.kt
@@ -43,9 +43,9 @@ data class Data(
     @SerializedName("rating")
     val rating: String,
     @SerializedName("score")
-    val score: Int,
+    val score: Float,
     @SerializedName("scored_by")
-    val scoredBy: Int,
+    val scoredBy: Float,
     @SerializedName("season")
     val season: String,
     @SerializedName("source")
